include files while booting app.
in some situations, files of the module need to access to config of the module, but right now files include in the regsiter method when config of modules doesn’t load yet ( it's will be available only if Laravel config is cached)

So, I think included files is best to load on boot function to ensure config of modules will be available.

any suggestion is welcome.