use default stub if override not exists
If somebody overrides stub path in the config, should copy new stubs with every new release of the package to add missing stub,

This PR fixes this problem by check if stub file exists on override path use it or use default stub form package folder as a fallback.

I add some test unit for that to make sure everything is work.

thanks